feat: smart auto-checkpoint before compaction (#4)
- Extract checkpoint logic into shared lib/checkpoint-writer.ts
- Refactor checkpoint tool to use shared helper (no behavior change)
- Add auto_checkpoint param to check_session_health (default: true)
- When session health is critical + dirty files exist, automatically
  write checkpoint and commit tracked files
- Closes #4
